TPT in MaPLe

+++ clip/maple_model.py
@@ -235,7 +235,7 @@ class ResidualAttentionBlock_IVLP(nn.Module):
                 # Remove the outputs produced by learnable tokens of previous layer
                 prefix = x[0:x.shape[0] - self.n_ctx_visual, :, :]
                 # Create/configure learnable tokens of this layer
-                visual_context = self.VPT_shallow.expand(x.shape[1], -1, -1).permute(1, 0, 2).half()
+                visual_context = self.VPT_shallow.expand(x.shape[1], -1, -1).permute(1, 0, 2)
                 # Add the learnable tokens of this layer with the input, by replacing the previous
                 # layer learnable tokens
                 x = torch.cat([prefix, visual_context], dim=0)
@@ -246,7 +246,7 @@ class ResidualAttentionBlock_IVLP(nn.Module):
                 prefix = x[:1, :, :]
                 suffix = x[1 + self.n_ctx_text:, :, :]
                 # Create/configure learnable tokens of this layer
-                textual_context = self.VPT_shallow.expand(x.shape[1], -1, -1).permute(1, 0, 2).half()
+                textual_context = self.VPT_shallow.expand(x.shape[1], -1, -1).permute(1, 0, 2)
                 # Add the learnable tokens of this layer with the input, replaced by previous
                 # layer learnable tokens
                 x = torch.cat([prefix, textual_context, suffix], dim=0)
@@ -303,7 +303,7 @@ class ResidualAttentionBlock_MaPLe(nn.Module):
                         prefix = x[0:x.shape[0] - self.compound_prompt_nctx, :, :]
                         # Create/configure learnable tokens of this layer
                         visual_context = compound_prompts_deeper[counter]  # extract the correct index
-                        visual_context = visual_context.expand(x.shape[1], -1, -1).permute(1, 0, 2).half()
+                        visual_context = visual_context.expand(x.shape[1], -1, -1).permute(1, 0, 2)
                         # Add the learnable tokens of this layer with the input, by replacing previous
                         # layer learnable tokens
                         x = torch.cat([prefix, visual_context], dim=0)
@@ -320,7 +320,7 @@ class ResidualAttentionBlock_MaPLe(nn.Module):
                         suffix = x[1 + self.compound_prompt_nctx:, :, :]
                         # Create/configure learnable tokens of this layer
                         textual_context = compound_prompts_deeper[counter]
-                        textual_context = textual_context.expand(x.shape[1], -1, -1).permute(1, 0, 2).half()
+                        textual_context = textual_context.expand(x.shape[1], -1, -1).permute(1, 0, 2)
                         # Add the learnable tokens of this layer with the input, replaced by previous
                         # layer learnable tokens
                         x = torch.cat([prefix, textual_context, suffix], dim=0)
@@ -402,7 +402,7 @@ class VisionTransformer(nn.Module):
         # After positional embeddings, we will attach prompts with the model, remember only those
         # are trainable parameters here in whole image encoder.
         if self.VPT_shallow:
-            visual_ctx = self.VPT.expand(x.shape[0], -1, -1).half()
+            visual_ctx = self.VPT.expand(x.shape[0], -1, -1)
             x = torch.cat([x, visual_ctx], dim=1)
         else:
             assert self.prompt_till_layer_visual == 0
@@ -454,7 +454,7 @@ class VisionTransformer_MaPLe(nn.Module):
         # After positional embeddings, we will attach prompts with the model, remember only those
         # are trainable parameters here in whole image encoder.
         if self.VPT_shallow:
-            visual_ctx = shared_ctx.expand(x.shape[0], -1, -1).half()
+            visual_ctx = shared_ctx.expand(x.shape[0], -1, -1)
             x = torch.cat([x, visual_ctx], dim=1)
         else:
             assert self.prompt_till_layer_visual == 0
